refactor(routing): split FeedsPage into AppLayout + nested child routes
Finding #3 from the audit. The 459-line FeedsPage component was
mounted by six routes (/feeds, /feeds/:feedId, /feeds/.../articles/...,
/explore, /stats, /settings) and branched on \`pathname\` to decide
which surface to render. Every new full-page surface (Stats,
Settings) bolted another \`isXxxPage\` flag into the same file.

Replaced with a router-driven shape that ADR 013 wanted in the
first place:

  src/pages/app-layout.tsx   — the shell. Sidebar on desktop,
                                header + drawer on mobile. Hosts
                                useKeyboardNav + the bare-/feeds
                                redirect. <Outlet /> for the stage.
  src/pages/feeds-route.tsx  — the feeds surface only: article list
                                + reader. Owns scroll-snap, article
                                sync, auto-select. Renders the
                                catalog inline when feeds=[] (matches
                                old behavior of "show explore in the
                                stage when there's nothing to read").
  src/pages/stage-view.tsx   — small scroll wrapper used by the
                                explore/stats/settings routes.

src/app.tsx now mounts each surface as a sibling child of <AppLayout>.
Adding a new full-page surface is "create a route + write a route
component," not "add another isXxxPage flag to the shell."

The stable [sidebar | stage] outer panel topology is preserved:
react-resizable-panels sees the same panel children on every route,
because <Outlet /> is one stable React node — exactly what ADR 013
demands.

Test updates: three test files (feeds-page-behavior, feeds-page-layout,
stats-route) now render the layout-route nesting. Old FeedsPage deleted.
Full suite stays 2487 passing, 0 type errors.
