+++ src/app.tsx
@@ -1,5 +1,5 @@
-import { useEffect, useState } from "react";
-import { BrowserRouter, Routes, Route } from "react-router";
+import { useEffect, useState, lazy, Suspense } from "react";
+import { BrowserRouter, Routes, Route, useNavigate } from "react-router";
 import { useAppStore } from "@/stores/app-store.ts";
 import { useFeedStore } from "@/stores/feed-store.ts";
 import { useArticleStore } from "@/stores/article-store.ts";
@@ -11,7 +11,9 @@ import { SyncMigrationDialog } from "@/components/sync/sync-migration-dialog.tsx
 import { DeviceSetupWizard } from "@/components/billing/device-setup-wizard.tsx";
 import { NavigateWithSearch } from "@/components/routing/navigate-with-search.tsx";
 import { SpeedInsights } from "@vercel/speed-insights/react";
-import { FeedsPage } from "@/pages/feeds-page.tsx";
+import { AppLayout } from "@/pages/app-layout.tsx";
+import { FeedsRoute } from "@/pages/feeds-route.tsx";
+import { StageView } from "@/pages/stage-view.tsx";
 import { BillingSuccess } from "@/pages/billing-success.tsx";
 import { BillingCancelled } from "@/pages/billing-cancelled.tsx";
 import { BillingRecover } from "@/pages/billing-recover.tsx";
@@ -24,6 +26,53 @@ import {
   type SecureContextProblemKind,
 } from "@/core/security/secure-context.ts";
 import { InvalidKeysScreen } from "@/components/recovery/invalid-keys-screen";
+
+const ExploreCatalog = lazy(() =>
+  import("@/components/explore/explore-catalog.tsx").then((m) => ({
+    default: m.ExploreCatalog,
+  })),
+);
+const StatsPage = lazy(() =>
+  import("@/components/stats/stats-page.tsx").then((m) => ({
+    default: m.StatsPage,
+  })),
+);
+const SettingsPage = lazy(() =>
+  import("@/pages/settings-page.tsx").then((m) => ({
+    default: m.SettingsPage,
+  })),
+);
+
+function ExploreRoute() {
+  const navigate = useNavigate();
+  return (
+    <StageView>
+      <Suspense>
+        <ExploreCatalog onFeedAdded={(id) => navigate(`/feeds/${id}`)} />
+      </Suspense>
+    </StageView>
+  );
+}
+
+function StatsRoute() {
+  return (
+    <StageView>
+      <Suspense>
+        <StatsPage />
+      </Suspense>
+    </StageView>
+  );
+}
+
+function SettingsRoute() {
+  return (
+    <StageView>
+      <Suspense>
+        <SettingsPage />
+      </Suspense>
+    </StageView>
+  );
+}
 
 function AppInit({ children }: { children: React.ReactNode }) {
   const isDbReady = useAppStore((s) => s.isDbReady);
@@ -220,15 +269,17 @@ export function App() {
       <BrowserRouter>
         <AppInit>
           <Routes>
-            <Route path="/feeds" element={<FeedsPage />} />
-            <Route path="/feeds/:feedId" element={<FeedsPage />} />
-            <Route
-              path="/feeds/:feedId/articles/:articleId"
-              element={<FeedsPage />}
-            />
-            <Route path="/explore" element={<FeedsPage />} />
-            <Route path="/stats" element={<FeedsPage />} />
-            <Route path="/settings" element={<FeedsPage />} />
+            <Route element={<AppLayout />}>
+              <Route path="/feeds" element={<FeedsRoute />} />
+              <Route path="/feeds/:feedId" element={<FeedsRoute />} />
+              <Route
+                path="/feeds/:feedId/articles/:articleId"
+                element={<FeedsRoute />}
+              />
+              <Route path="/explore" element={<ExploreRoute />} />
+              <Route path="/stats" element={<StatsRoute />} />
+              <Route path="/settings" element={<SettingsRoute />} />
+            </Route>
             <Route path="/billing/success" element={<BillingSuccess />} />
             <Route path="/billing/cancelled" element={<BillingCancelled />} />
             <Route path="/billing/recover" element={<BillingRecover />} />
